requirements.txt
# ── Core RAG Stack ──────────────────────────────────────────
groq>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# ── PDF Processing for Research Papers ──────────────────────
pypdf>=3.0.0
pdfplumber>=0.10.0  

# ── Document Processing ─────────────────────────────────────
python-docx>=0.8.11
tiktoken>=0.4.0

# ── Ground Truth / Evaluation ───────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ── UI ──────────────────────────────────────────────────────
streamlit>=1.28.0
plotly>=5.14.0

# ── Utilities ───────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
tenacity>=8.2.0
rich>=13.0.0

# ── Testing ─────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0